fix: update tests & snapshots
